+++ src/ImageResizer.Plugins.AutoCrop/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace ImageResizer.Plugins.AutoCrop
+{
+    public static class Constants
+    {
+        public static readonly double BytePrecision = 1 / (double)byte.MaxValue;
+    }
+}
