+++ src/ImageResizer.Plugins.AutoCrop/Extensions/ColorExtensions.cs
@@ -6,7 +6,6 @@ namespace ImageResizer.Plugins.AutoCrop.Extensions
     public static class ColorExtensions
     {
         private static readonly byte _maxBuckets = 20;
-        private static readonly double _bytePrecision = 1 / (double)byte.MaxValue;
         private static readonly double _bucketPrecision = (_maxBuckets + 1) / (double)byte.MaxValue;
 
         public static byte ToColorBucket(this Color color)
@@ -18,7 +17,7 @@ namespace ImageResizer.Plugins.AutoCrop.Extensions
         {
             var mix = (uint)(0.299 * color.R + 0.587 * color.G + 0.114 * color.B);
 
-            return (byte)Math.Min(byte.MaxValue, mix * (color.A * _bytePrecision) + (byte.MaxValue - color.A));
+            return (byte)Math.Min(byte.MaxValue, mix * (color.A * Constants.BytePrecision) + (byte.MaxValue - color.A));
         }
 
         public static byte GetMaxColorBuckets()
