+++ src/ImageResizer.Plugins.AutoCrop/ImageResizer.Plugins.AutoCrop.csproj
@@ -49,10 +49,11 @@
     <Reference Include="System.Drawing" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Actions\RawCopy.cs" />
+    <Compile Include="Actions\Raw.cs" />
     <Compile Include="Analyzers\BorderAnalyzer.cs" />
     <Compile Include="Analyzers\BoundsAnalyzer.cs" />
     <Compile Include="AutoCropPlugin.cs" />
+    <Compile Include="Constants.cs" />
     <Compile Include="Extensions\BitmapExtensions.cs" />
     <Compile Include="Extensions\ColorExtensions.cs" />
     <Compile Include="Extensions\PointExtensions.cs" />
